requirements.txt
# Rasa Pro and Core
rasa[full]==3.6.0
rasa-pro

# AI and Machine Learning
google-generativeai>=0.3.0
google-cloud-speech>=2.20.0
google-cloud-texttospeech>=2.14.0
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment if you have GPU support

# Natural Language Processing
nltk>=3.8.1

# Data Science and Analytics
pandas>=2.1.0
numpy>=1.24.0

# Web Framework and APIs
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# Logging and Utilities
typing-extensions>=4.5.0

# Development and Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Additional utilities you might need
python-dateutil>=2.8.2
pytz>=2023.3
uuid>=1.30

# For better JSON handling
ujson>=5.8.0

# HTTP client improvements
httpx>=0.24.0

# For better regex support
regex>=2023.6.3

# Async support
asyncio>=3.4.3

# For better traceback handling
traceback2>=1.4.0

# Pickle security (if needed)
dill>=0.3.7

# Memory optimization
psutil>=5.9.0

# Optional: Database support
# psycopg2-binary>=2.9.0  # For PostgreSQL
# sqlalchemy>=2.0.0       # For ORM

# Optional: Monitoring and metrics
# prometheus-client>=0.17.0

# Optional: Additional ML libraries
 scikit-learn>=1.3.0
 transformers>=4.30.0